# Faster Load(array)