--- a/source/Nevermore/CommandParameterValues.cs
+++ b/source/Nevermore/CommandParameterValues.cs
@@ -2,10 +2,13 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Data;
+using System.Data.Common;
 #if NETFRAMEWORK
 using System.Data.SqlClient;
+using Microsoft.SqlServer.Server;
 #else
 using Microsoft.Data.SqlClient;
+using Microsoft.Data.SqlClient.Server;
 #endif
 using System.Linq;
 using System.Reflection;
@@ -50,6 +53,18 @@ namespace Nevermore
 
         public CommandType CommandType { get; set; }
 
+        public void AddTable(string name, IEnumerable<string> ids)
+        {
+            var idColumnMetadata = new SqlMetaData("ParameterValue", SqlDbType.NVarChar, 300);
+            
+            Add(name, ids.Where(v => !string.IsNullOrWhiteSpace(v)).Select(v =>
+            {
+                var record = new SqlDataRecord(idColumnMetadata);
+                record.SetValue(0, v);
+                return record;
+            }).ToList());
+        }
+        
         void AddFromParametersObject(object args)
         {
             if (args == null)
@@ -79,6 +94,14 @@ namespace Nevermore
             if (value == null)
             {
                 command.Parameters.Add(new SqlParameter(name, DBNull.Value));
+                return;
+            }
+
+            if (value is List<SqlDataRecord> dr && command is SqlCommand sqlCommand)
+            {
+                var p = sqlCommand.Parameters.Add(name, SqlDbType.Structured);
+                p.Value = dr;
+                p.TypeName = "dbo.[ParameterList]";
                 return;
             }
 
@@ -137,6 +160,141 @@ namespace Nevermore
                     throw new Exception($"The parameter {item.Key} already exists");
                 this[item.Key] = item.Value;
             }
+        }
+    }
+
+    public class MyDataReader : DbDataReader
+    {
+        public override bool GetBoolean(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override byte GetByte(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override long GetBytes(int ordinal, long dataOffset, byte[] buffer, int bufferOffset, int length)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override char GetChar(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override long GetChars(int ordinal, long dataOffset, char[] buffer, int bufferOffset, int length)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override string GetDataTypeName(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override DateTime GetDateTime(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override decimal GetDecimal(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override double GetDouble(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override Type GetFieldType(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override float GetFloat(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override Guid GetGuid(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override short GetInt16(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int GetInt32(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override long GetInt64(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override string GetName(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int GetOrdinal(string name)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override string GetString(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override object GetValue(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int GetValues(object[] values)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override bool IsDBNull(int ordinal)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int FieldCount { get; }
+
+        public override object this[int ordinal] => throw new NotImplementedException();
+
+        public override object this[string name] => throw new NotImplementedException();
+
+        public override int RecordsAffected { get; }
+        public override bool HasRows { get; }
+        public override bool IsClosed { get; }
+
+        public override bool NextResult()
+        {
+            throw new NotImplementedException();
+        }
+
+        public override bool Read()
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int Depth { get; }
+
+        public override IEnumerator GetEnumerator()
+        {
+            throw new NotImplementedException();
         }
     }
 }