Nevermore provides a Load overload that takes an array of ID's. The common use case is that you perform a select, find a set of related document IDs, then load that set of related documents. 

It currently works by generating large "IN" statements. E.g.,:

    SELECT * FROM Product WHERE Id in (@ids_0, @ids_1, .....)

This has a few downsides:

- You can only have so many parameters in a statement, so when searching on hundreds of ID's, it needs to split them into chunks
 - Queries with different numbers of parameters (e.g., 317 vs 407) don't get to reuse the same query plan
 - It's generally just slower

This new version makes use of tabled value parameters. 

Here's the performance difference (on my local SQLEXPRESS)

Old approach:

```
Loaded 100 products in 42ms  // warm up
Loaded 100 products in 2ms 
Loaded 300 products in 32ms
Loaded 600 products in 79ms
Loaded 1200 products in 105ms
Loaded 1800 products in 130ms
Loaded 2400 products in 166ms
```

New approach (table valued parameters)

```
Loaded 100 products in 47ms
Loaded 100 products in 2ms
Loaded 300 products in 5ms
Loaded 600 products in 9ms
Loaded 1200 products in 16ms
Loaded 1800 products in 23ms
Loaded 2400 products in 30ms
```

The new approach is faster because SQL is able to reuse the same query plan whether you ask for 5 ID's or 5000. It also doesn't have to batch, so it can load thousands in one go. 

For large data sets, the difference is more obvious. Here it is loading 50,000 items:

```
Loaded 50000 products in 684ms (new way)
Loaded 50000 products in 7347ms (old way)
```

This feels fast enough (for 50,000 items!) that it might move this from being a strategy we want to avoid to one we can actually use more frequently. 

**There is one catch though**. You have to define this type (with this name):

```
CREATE TYPE dbo.[ParameterList] as TABLE([ParameterValue] nvarchar(300))
```

This would need to be a migration script in Octopus/Octofront.