@PaulStovell 2020-04-06
A bit more about why it seems to be faster.

The giant "IN" clause generates a query plan like this:

![image](https://user-images.githubusercontent.com/47085/78514924-28998a80-77f7-11ea-8458-8f351b3bd258.png)

While the table valued parameter approach looks like this:

![image](https://user-images.githubusercontent.com/47085/78514939-36e7a680-77f7-11ea-9c69-70980d96e171.png)

The TVP approach only scans the parameter list - not the source table (which would likely be much larger).