@tothegills 2020-04-06
[Table value parameters reference](https://docs.microsoft.com/en-us/sql/relational-databases/tables/use-table-valued-parameters-database-engine?view=sql-server-ver15)